# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/me": {
      "get": {
        "operationId": "UsersController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Current user profile and status",
        "tags": [
          "me"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change display name",
        "tags": [
          "me"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete (anonymise) the account",
        "tags": [
          "me"
        ]
      }
    },
    "/verification/platform": {
      "post": {
        "operationId": "VerificationController_platform",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlatformAgeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Record the operating system's age-range answer (Apple Declared Age Range / Play Age Signals)",
        "tags": [
          "verification"
        ]
      }
    },
    "/verification/status": {
      "get": {
        "operationId": "VerificationController_status",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Current age assurance state",
        "tags": [
          "verification"
        ]
      }
    },
    "/venues/nearby": {
      "get": {
        "operationId": "VenuesController_nearby",
        "parameters": [
          {
            "name": "lat",
            "required": true,
            "in": "query",
            "schema": {
              "example": 52.5219,
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": true,
            "in": "query",
            "schema": {
              "example": 13.4132,
              "type": "number"
            }
          },
          {
            "name": "accuracy",
            "required": true,
            "in": "query",
            "description": "Horizontal accuracy radius in metres (68% confidence)",
            "schema": {
              "example": 12,
              "type": "number"
            }
          },
          {
            "name": "mocked",
            "required": false,
            "in": "query",
            "description": "True when the OS reports a mock/simulated location",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "viewLat",
            "required": false,
            "in": "query",
            "description": "Latitude of the map viewport centre; defaults to the fix",
            "schema": {
              "example": 48.1374,
              "type": "number"
            }
          },
          {
            "name": "viewLng",
            "required": false,
            "in": "query",
            "description": "Longitude of the map viewport centre; defaults to the fix",
            "schema": {
              "example": 11.5755,
              "type": "number"
            }
          },
          {
            "name": "viewRadiusM",
            "required": false,
            "in": "query",
            "description": "Half the visible map width in metres; capped at DISCOVERY_MAX_RADIUS_M. Cafés are only included up to DISCOVERY_RADIUS_M.",
            "schema": {
              "example": 1500,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Real places where the map is looking (discovered from OpenStreetMap), nearest to the phone first",
        "tags": [
          "venues"
        ]
      }
    },
    "/venues/{id}": {
      "get": {
        "operationId": "VenuesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Basic venue info",
        "tags": [
          "venues"
        ]
      }
    },
...
  },
  "info": {
    "title": "Larea API",
    "description": "Location-based chat backend. Realtime events are documented in docs/realtime-protocol.md.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateMeDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "description": "Display name must be 3-24 characters: letters, numbers, spaces, underscores or dots."
          }
        },
        "required": [
          "displayName"
        ]
      },
      "PlatformAgeDto": {
        "type": "object",
        "properties": {
          "platform": {
            "type": "string",
            "enum": [
              "apple",
              "google"
            ]
          },
          "lowerBound": {
            "type": "number",
            "description": "Lower bound of the shared age range; absent when the user declined"
          },
          "upperBound": {
            "type": "number"
          },
          "declaration": {
            "type": "string",
            "enum": [
              "self",
              "guardian",
              "confirmed",
              "unknown"
            ],
            "description": "How the platform obtained the age range"
          }
        },
        "required": [
          "platform",
          "declaration"
        ]
      },
...
    }
  }
}